chore: feedback amends
